Solve the code challenge 1 from course 5 by adding the necessary files and implementing the code.
