+++ app/src/main/AndroidManifest.xml
@@ -11,8 +11,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.AndroidFundamentalsMai2021">
-        <activity android:name=".activities.SecondActivity"></activity>
-        <activity android:name=".activities.FirstActivity" />
+        <activity android:name=".course5codechallenge1.SecondActivity" />
+        <activity android:name=".course5codechallenge1.FirstActivity" />
         <activity android:name=".ConstraintLayoutSampleActivity" />
         <activity android:name=".MainActivity" />
         <activity android:name=".EmailsActivity" />
